fix(P1): Float→Decimal 精度修复 — 7 个核心文件 + 31 个新增精度测试
根因：25+ 处 float() 转换分布在 10+ 个文件中，部分污染了决策链
(score 比较、pnl_ratio 累积、回测报告精度)。

修复 6 个核心文件：
- strategy_engine.py: calculate_score() 返回类型 float → Decimal
- models.py: PatternResult.score → Decimal, SignalResult.pnl_ratio → Optional[Decimal]
- backtester.py: _simulate_win_rate 和 _calculate_attempt_outcome 返回值 → Decimal
- engulfing_strategy.py: score 保持 Decimal, 边界比较改用 Decimal
- risk_calculator.py: score 参数类型 float → Decimal
- signal_pipeline.py: _check_cover score 参数 → Decimal

新增 31 个精度测试（23 单元 + 7 集成 + 1 预存修复），全部通过。
P2 保留：details={} 字典中 float() 不变（仅 JSON 序列化用途）。

+++ src/application/signal_pipeline.py
@@ -881,7 +881,7 @@ class SignalPipeline:
         self,
         kline: KlineData,
         attempt: SignalAttempt,
-        score: float,
+        score: Decimal,
     ) -> Tuple[bool, Optional[str], Optional[Dict[str, Any]]]:
         """
         Check if new signal should cover (supersede) existing active signal.
